0010606: In workflow,delete item in MDM couldn't produce a approval process automatically.
git-svn-id: http://talendforge.org/svn/mdm/trunk@34308 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.webapp.itemsbrowser/src/com/amalto/webapp/v3/itemsbrowser/dwr/ItemsBrowserDWR.java
@@ -1067,6 +1067,11 @@ public class ItemsBrowserDWR {
 		try {
 			Configuration config = Configuration.getInstance();
 			String dataClusterPK = config.getCluster();
+			
+			if(com.amalto.core.util.Util.beforeDeleting(dataClusterPK,concept,ids)){
+				return "OK - But go through the beforeDeleting transformer first";
+			}
+			
 			TreeNode rootNode = getRootNode(concept, "en");
 	        if(ids!=null && !rootNode.isReadOnly()){
 				WSItemPK wsItem = Util.getPort().deleteItem(
@@ -1087,8 +1092,9 @@ public class ItemsBrowserDWR {
 		}
 		catch(Exception e){
 			return "ERROR -" + e.getLocalizedMessage();
-		}        
+		}       
 	}
+	
 	public String[] getUriArray(String concept, String[] ids){
 		Configuration config;
 		List<String> uriList=new ArrayList<String>();
